fix: movve url request
